First lint pass.
