+++ .config/beemo.ts
@@ -1,16 +1,19 @@
 export default {
-  module: '@beemo/dev',
-  drivers: [
-    ['babel', { strategy: 'none' }],
-    'eslint',
-    'jest',
-    'prettier',
-    [
-      'typescript',
-      {
-        buildFolder: 'dts',
-        declarationOnly: true,
-      },
-    ],
-  ],
+	module: '@beemo/dev',
+	drivers: [
+		['babel', { strategy: 'none' }],
+		'eslint',
+		'jest',
+		'prettier',
+		[
+			'typescript',
+			{
+				buildFolder: 'dts',
+				declarationOnly: true,
+			},
+		],
+	],
+	settings: {
+		react: true,
+	},
 };
