+++ packages/core-react/src/createDirectionHelpers.tsx
@@ -1,74 +1,75 @@
 import React, { createContext, useContext } from 'react';
 import getDirection from 'direction';
 import { Aesthetic, Direction } from '@aesthetic/core';
-import createHOC from './createHOC';
+import { createHOC } from './createHOC';
 import {
-  DirectionContextType,
-  DirectionProviderProps,
-  WithDirectionWrappedProps,
-  WrapperComponent,
-  WrapperProps,
+	DirectionContextType,
+	DirectionProviderProps,
+	WithDirectionWrappedProps,
+	WrapperComponent,
+	WrapperProps,
 } from './types';
 
-interface DirectionHelperOptions {
-  wrapper: React.ElementType;
+export interface DirectionHelperOptions {
+	wrapper: React.ElementType;
 }
 
-export default function createDirectionHelpers<Result, Block extends object>(
-  aesthetic: Aesthetic<Result, Block>,
-  { wrapper: Wrapper }: DirectionHelperOptions,
+export function createDirectionHelpers<Result, Block extends object>(
+	aesthetic: Aesthetic<Result, Block>,
+	{ wrapper: Wrapper }: DirectionHelperOptions,
 ) /* infer */ {
-  const DirectionContext = createContext<DirectionContextType>('ltr');
+	const DirectionContext = createContext<DirectionContextType>('ltr');
 
-  /**
-   * Explicitly define a direction or automatically infer a direction based on a string of text.
-   * Will render an element with a `dir` attribute set.
-   */
-  function DirectionProvider({ children, direction, value }: DirectionProviderProps) {
-    const dir = (direction ||
-      (value && getDirection(value)) ||
-      aesthetic.getActiveDirection() ||
-      'ltr') as Direction;
+	/**
+	 * Explicitly define a direction or automatically infer a direction based on a string of text.
+	 * Will render an element with a `dir` attribute set.
+	 */
+	function DirectionProvider({ children, direction, value }: DirectionProviderProps) {
+		const dir = (direction ||
+			(value && getDirection(value)) ||
+			aesthetic.getActiveDirection() ||
+			'ltr') as Direction;
 
-    return (
-      <DirectionContext.Provider value={dir}>
-        <Wrapper dir={dir}>{children}</Wrapper>
-      </DirectionContext.Provider>
-    );
-  }
+		return (
+			<DirectionContext.Provider value={dir}>
+				<Wrapper dir={dir}>{children}</Wrapper>
+			</DirectionContext.Provider>
+		);
+	}
 
-  /**
-   * Hook within a component to return the current RTL direction.
-   */
-  function useDirection(): Direction {
-    return useContext(DirectionContext);
-  }
+	/**
+	 * Hook within a component to return the current RTL direction.
+	 */
+	function useDirection(): Direction {
+		return useContext(DirectionContext);
+	}
 
-  /**
-   * Wrap a React component with an HOC that injects the current direction object as a prop.
-   */
-  function withDirection() /* infer */ {
-    return function withDirectionComposer<Props extends object = {}>(
-      WrappedComponent: React.ComponentType<Props & WithDirectionWrappedProps>,
-    ): React.FunctionComponent<Omit<Props, keyof WithDirectionWrappedProps> & WrapperProps> &
-      WrapperComponent {
-      return createHOC(
-        'withDirection',
-        WrappedComponent,
-        // eslint-disable-next-line prefer-arrow-callback
-        function WithDirection({ wrappedRef, ...props }) {
-          const direction = useDirection();
+	/**
+	 * Wrap a React component with an HOC that injects the current direction object as a prop.
+	 */
+	function withDirection() /* infer */ {
+		return function withDirectionComposer<Props extends object = {}>(
+			WrappedComponent: React.ComponentType<Props & WithDirectionWrappedProps>,
+		): React.FunctionComponent<Omit<Props, keyof WithDirectionWrappedProps> & WrapperProps> &
+			WrapperComponent {
+			return createHOC(
+				'withDirection',
+				WrappedComponent,
+				// eslint-disable-next-line prefer-arrow-callback
+				function WithDirection({ wrappedRef, ...props }) {
+					const direction = useDirection();
 
-          return <WrappedComponent {...(props as any)} ref={wrappedRef} direction={direction} />;
-        },
-      );
-    };
-  }
+					// eslint-disable-next-line @typescript-eslint/no-explicit-any
+					return <WrappedComponent {...(props as any)} ref={wrappedRef} direction={direction} />;
+				},
+			);
+		};
+	}
 
-  return {
-    DirectionContext,
-    DirectionProvider,
-    useDirection,
-    withDirection,
-  };
+	return {
+		DirectionContext,
+		DirectionProvider,
+		useDirection,
+		withDirection,
+	};
 }
