+++ packages/react/src/aesthetic.ts
@@ -2,11 +2,9 @@ import { Aesthetic, ClassName, ElementStyles } from '@aesthetic/core';
 import { createClientEngine } from '@aesthetic/style';
 import { isDOM } from '@aesthetic/utils';
 
-const aesthetic = new Aesthetic<ClassName, ElementStyles>();
+export const aesthetic = new Aesthetic<ClassName, ElementStyles>();
 
 // Avoid creating the client on the server
 if (isDOM()) {
-  aesthetic.configureEngine(createClientEngine());
+	aesthetic.configureEngine(createClientEngine());
 }
-
-export default aesthetic;
