+++ packages/react/src/direction.ts
@@ -1,9 +1,9 @@
 import { createDirectionHelpers } from '@aesthetic/core-react';
-import aesthetic from './aesthetic';
+import { aesthetic } from './aesthetic';
 
 export const {
-  DirectionContext,
-  DirectionProvider,
-  useDirection,
-  withDirection,
+	DirectionContext,
+	DirectionProvider,
+	useDirection,
+	withDirection,
 } = createDirectionHelpers(aesthetic, { wrapper: 'div' });
